refactor(core): combine atob + btoa modules
